Remove extra TradeData class
